update dict when interactively spellchecking
Before if we interactively added words to the dictionary they were not
persisted since they were written to a temporary file.
Then you need to add the words manually to the dictionary which is
annoying and also error prone since it should remain sorted.

outdated comment, we use `aspell` with `--dont-backup`