Chapter 4 (#30)
* chore: dump chapter-4 drafts

* split out nil section

change the constructed example to use the Color theme instead

* edit type assertion

- use standalone example file
- fix missing dependencies
- using simple Color interface without full specs
- add relevant links
- include full section example

* edit Abstracting memory and interfaces

- adapt to image interface changes
- grammar

* edit full image example

* edit comparable

- grammar

* edit section implements

gramar etc.

* edit termination

With warning that this is not the current Gobra behavior

* add TODO note for expected changes

* add subtypes to dict

* replace bit operations; drop assume

we include bounds for pix elements in the invariant for Alpha16Image
since this is not inferred from the type.

now we can prove that the color is valid

* remove failing examples from full section example

it might be confusable if the part intended to copy-paste causes 4
errors even if they are intended

---------

Co-authored-by: João Pereira <joaopereira.19@gmail.com>