Gobra book addons (#32)
* Add skeleton summary

Included a test page to see that the deployments work as expected.

* Add simple gobra support to ace editor

* Add quiz support

Via the preprocessor https://github.com/cognitive-engineering-lab/mdbook-quiz
Added a testquiz to the testpage

* Draft introduction contents

* Improve highlighting // @ comments

* Write out first part on basic specs

Using newtons method example

* Test images

* Add favicon

Exported the png version from the svg in inkscape (64x64)

* Grammar

* Remove testpage

We can use for local testing but it should not be included in the repo.

* Remove contribution related information

to be added to CONTRIBUTING.md

* Revert title

* Remove content not yet ready for review

* Hide lines

* readonly/editable code blocks

* add first hljs highlighting for gobra

* feat: verify button

* add debug page

* fix: favicon

* chore: fmt

* hidden buttons

* add: dump spec hiding (not working)

* chore: fmt

* Highlight Specs (instead of hiding)

* highlight on hiding button

* fix: //@  may come after some text

* chore: fmt index.hbs

* fix: prevent loading editor.js twice

* feat: ace hiding lines

without highlightjs as previously done

had to remove the character from the book.toml
as this preprocessed the code

now we construct a hidden and non-hidden version
of the code and toggle between theme

note this makes only sense in readonly mode
writing and interfering with hidden code
sounds like a nightmare

* chore: init typescript

* refacotr: global context map with uuids

* fix!: include as module

change ts outputs

this could be dangerous if we try to include other custom js...

* chore: ignore emitted js

* chore: type it

* refactor: hidden button

* feat: add simpleButton

* refactor: clipboard button

* refactor: verify/run buttons

* refactor: undo/reset button

* refactor: book

* chore: comments and details

* match // @ also with whitespace

* fix: consistent theme

the theme would not properly be set up on first page load
after changing the theme on another page the theme editors on a new page
would still use the old theme

* fix: only show hide when there is hidden code

* only initialize editor for go and gobra

previously an error block (with language "text")
got an editor as well which made it look weird

* chore: remove merge artifacts

* playground page

* revert typescript

did not want to update the build action again
also don't want to commit both ts and generated js

* update contributing infos

* no abstract function here

would not be valid go

* show line number for editable files

* fix error (text) block display

* fix: remove (one) trailing line displayed for all code blocks

* fix: code indentation

* fix: wrong language

* chore: remove old files